get projects count for organization

+++ Src/Libraries/2-Application/Application.Core/Dtos/Team/Organizations/OrganizationOutputDto.cs
@@ -6,7 +6,7 @@ namespace TaskoMask.Application.Core.Dtos.Team.Organizations
     public class OrganizationOutputDto : OrganizationBasicInfoDto
     {
         [Display(Name = nameof(ApplicationMetadata.ProjectsCount), ResourceType = typeof(ApplicationMetadata))]
-        public int ProjectsCount { get; set; }
+        public long ProjectsCount { get; set; }
         [Display(Name = nameof(ApplicationMetadata.OwnerDisplayName), ResourceType = typeof(ApplicationMetadata))]
         public string OwnerMemberDisplayName { get; set; }
     }
